# Various performance improvements in the `insertRows` path