--- a/src/main/java/net/snowflake/ingest/streaming/internal/MemoryInfoProviderFromRuntime.java
+++ b/src/main/java/net/snowflake/ingest/streaming/internal/MemoryInfoProviderFromRuntime.java
@@ -13,8 +13,10 @@ public class MemoryInfoProviderFromRuntime implements MemoryInfoProvider {
   private final long maxMemory;
   private volatile long totalFreeMemory;
   private final ScheduledExecutorService executorService;
+  private static final long FREE_MEMORY_UPDATE_INTERVAL_MS = 1000;
+  private static final MemoryInfoProviderFromRuntime INSTANCE = new MemoryInfoProviderFromRuntime(FREE_MEMORY_UPDATE_INTERVAL_MS);
 
-  public MemoryInfoProviderFromRuntime(long freeMemoryUpdateIntervalMs) {
+  private MemoryInfoProviderFromRuntime(long freeMemoryUpdateIntervalMs) {
     maxMemory = Runtime.getRuntime().maxMemory();
     totalFreeMemory =
         Runtime.getRuntime().freeMemory() + (maxMemory - Runtime.getRuntime().totalMemory());
@@ -33,6 +35,10 @@ public class MemoryInfoProviderFromRuntime implements MemoryInfoProvider {
   private void updateFreeMemory() {
     totalFreeMemory =
         Runtime.getRuntime().freeMemory() + (maxMemory - Runtime.getRuntime().totalMemory());
+  }
+
+  public static MemoryInfoProviderFromRuntime getInstance() {
+    return INSTANCE;
   }
 
   @Override