--- a/src/test/java/net/snowflake/ingest/streaming/internal/SnowflakeStreamingIngestChannelTest.java
+++ b/src/test/java/net/snowflake/ingest/streaming/internal/SnowflakeStreamingIngestChannelTest.java
@@ -62,11 +62,6 @@ public class SnowflakeStreamingIngestChannelTest {
     }
 
     @Override
-    public long getTotalMemory() {
-      return maxMemory;
-    }
-
-    @Override
     public long getFreeMemory() {
       return freeMemory;
     }