… pom.xml
… src/test/java/net/snowflake/ingest/streaming/internal/SnowflakeStreamingIngestClientTest.java