@sfc-gh-tzhang 2024-06-25
> Overall, 35-40% improvement in the benchmark observed.

Wow, this is huge! Could you add in the describe what types of benchmark did you do? Is it targeted benchmarks for these cases or we're seeing this across all use cases?