@sfc-gh-psaha 2024-06-25
> > Overall, 35-40% improvement in the benchmark observed.
> 
> Wow, this is huge! Could you add in the describe what types of benchmark did you do? Is it targeted benchmarks for these cases or we're seeing this across all use cases?

I have the benchmark in this PR and now updated the PR description to describe the benchmark and where the impact will be seen mostly. I think there are probably other areas to improve in the data validation and also in the ParquetRowBuffer code but that's another fight for another day :)