@jdcaperon 2024-08-03
@sfc-gh-psaha if you would like real world samples as well for performance I might be able to provide some data from our production environment.